+++ ogma-cli/src/CLI/CommandROSApp.hs
@@ -236,7 +236,7 @@ commandOptsParser = CommandOpts
         <> metavar "FORMAT_NAME"
         <> help strROSAppFormatDesc
         <> showDefault
-        <> value "fcs"
+        <> value "default"
         )
   <*> strOption
         (  long "prop-format"